- Added the model.py file in solution and lab folder - Added Setup.py file in solution and lab folder -  Updated the `streaming_minute_traffic_pipeline.py` with TODO's part in lab folder - Updated the `streaming_minute_traffic_pipeline.py` in solution folder
