+++ quests/dataflow_python/7_Advanced_Streaming_Analytics/solution/model.py
@@ -1,0 +1,15 @@
+import typing
+import apache_beam as beam
+
+class CommonLog(typing.NamedTuple):
+    ip: str
+    user_id: str
+    lat: float
+    lng: float
+    timestamp: str
+    http_request: str
+    http_response: int
+    num_bytes: int
+    user_agent: str
+
+beam.coders.registry.register_coder(CommonLog, beam.coders.RowCoder)
