+++ quests/dataflow_python/7_Advanced_Streaming_Analytics/solution/setup.py
@@ -1,0 +1,8 @@
+import setuptools
+
+setuptools.setup(
+    name='my-pipeline',
+    version='1.0',
+    install_requires=[],
+    py_modules=['model'],
+)
